requirements.txt
Django==5.1.1
celery==5.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
django-redis==5.4.0
flower==2.0.1